refactor package structure token parts
