+++ src/main/java/digdir/dc24_eu_wallet/idTokens/TokenPayload.java
@@ -1,19 +1,20 @@
-package digdir.dc24_eu_wallet.aport.fromAnsattporten;
+package digdir.dc24_eu_wallet.idTokens;
 
 import com.google.gson.Gson;
 
 import java.io.UnsupportedEncodingException;
 import java.util.Base64;
 
+import digdir.dc24_eu_wallet.idTokens.ansattporten.fromAnsattporten.TokenHead;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.oauth2.core.oidc.OidcIdToken;
 
 /**
  * Responsible for decoding the Oidc id token of the user logged in to
- * ansattporten and returning it upon request.
+ * ansattporten or idporten, and returning it upon request.
  *
  * @Author Langbakk
- * @Version 10.07.2024
+ * @Version 23.07.2024
  */
 public class TokenPayload {
 
@@ -60,6 +61,12 @@ public class TokenPayload {
         return payload;
     }
 
+    /**
+     * Reads the token from ansattporten into an object.
+     *
+     * @param payload the payload part of the id token.
+     * @return object of ansattporten token payload part.
+     */
     public TokenHead getTokenHeadAnsattporten(String payload){
         //Using tool Gson, we will read the decoded token into classes created
         //for containing the information in the token.
@@ -78,8 +85,6 @@ public class TokenPayload {
      * id porten-
      *
      * @return
-
-
     public TokenHeadIdPorten getTokenHeadIdPOrten(){
         //Using tool Gson, we will read the decoded token into classes created
         //for containing the information in the token.
@@ -87,7 +92,7 @@ public class TokenPayload {
 
         //Parses from JSON into object. In this instance it will parse payload
         //into class TokenHead, which is the topmost class of our token.
-        tokenPayloadAsObject = gson.fromJson(payload, TokenHeadIdPorten.class);
+        tokenHeadIdPorten = gson.fromJson(payload, TokenHeadIdPorten.class);
 
         return tokenPayloadAsObject;
     }
